% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{css_dependency}
\alias{css_dependency}
\title{Create CSS HTML dependency}
\usage{
css_dependency(css_names = NULL)
}
\arguments{
\item{css_names}{Base names of CSS files to include, "common" always included}
}
\value{
An object that can be included in a list of dependencies passed to attachDependencies.
}
\description{
Create CSS HTML dependency
}
