% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_html.R
\name{html_worksheet}
\alias{html_worksheet}
\title{HTML format for literate, interactive R Markdown worksheets}
\usage{
html_worksheet(template = "default", ...)
}
\arguments{
\item{template}{User templates are not supported}

\item{...}{See other arguments in the documentation for \code{\link[rmarkdown:html_document]{rmarkdown::html_document()}}
(note that you cannot use the \code{theme} argument because it is set
internally).}
}
\value{
An R Markdown output format.
}
\description{
This format is for an HTML version of worksheets used in
my Quantitative Methods in Biological Anthropology course.
}
