% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_pdf.R
\name{pdf_worksheet}
\alias{pdf_worksheet}
\alias{pdf_assignment}
\title{PDF formats for literate, interactive R Markdown documents}
\usage{
pdf_worksheet(fig_width = 6, fig_height = 3.7, ...)

pdf_assignment(...)
}
\arguments{
\item{fig_width}{Default width (in inches) for figures}

\item{fig_height}{Default height (in inches) for figures}

\item{..., fig_crop, dev, df_print, highlight, keep_tex, keep_md, latex_engine, citation_package, includes, md_extensions, output_extensions, pandoc_args, extra_dependencies}{Other arguments to be passed to \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document()}}
(note you cannot use the \code{template} argument in either
\code{pdf_worksheet} or \code{pdf_assignment}; this argument is set
internally).}
}
\value{
An R Markdown output format.
}
\description{
These formats are for PDF versions of worksheets and assignments used in
my Quantitative Methods in Biological Anthropology course.
}
\section{Details}{
 You can find more details about each output format below.
}

\section{\code{pdf_worksheet}}{
 Format for creating a worksheet including fancy
page headers and logo on the first page.
}

\section{\code{pdf_assignment}}{
 Format for creating a minimally styled
assignment document.
}

